unit tests added. integration tests are done.

+++ src/main/java/com/cansuiremkanli/libmanage/service/impl/UserServiceImpl.java
@@ -50,6 +50,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException("User not found");
+        }
         userRepository.deleteById(id);
     }
+
 }
